pyproject.toml
[project]
name = "projeto-etl-satc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.0.0",
    "azure-core>=1.31.0",
    "azure-identity>=1.19.0",
    "azure-storage-file-datalake>=12.17.0",
    "delta-spark==3.2.0",
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "pymongo>=4.10.1",
    "pyodbc>=5.2.0",
    "pyspark==3.5.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.7.0",
    "sqlalchemy>=1.4.49,<2.0",
]
